backend/library/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
Faker==12.0.1
future==0.18.3
generics==6.0.0
Markdown==3.4.1
mixer==7.2.2
python-dateutil==2.8.2
pytz==2022.7
serializers==0.2.4
six==1.16.0
sqlparse==0.4.3
